OpenTelemetry replaces OpenCensus for tracing
If the OpenTelemetry SDK is started then tracing is now
enabled by default.

Span operations are no-op's in `opentelemetry_api` if no SDK
(the `opentelemetry` package) is started this doesn't change
anything for anyone not using OpenTelemetry.

`include_statement_span_attribute` option is added for
enabling/disabling the inclusion of the whole query as an
attribute on the OpenTelemetry Span started for a query.